improved usability

+++ app/src/main/java/com/santiagogil/takestock/controller/ConsumptionController.java
@@ -6,14 +6,14 @@ import com.santiagogil.takestock.model.daos.ConsumptionsDAO;
 
 public class ConsumptionController {
 
-    public void addConsumptionToDatabases(Context context, Integer itemID){
+    public void addConsumptionToDatabases(Context context, Long itemID){
 
         ConsumptionsDAO consumptionsDAO = new ConsumptionsDAO(context);
         consumptionsDAO.addConsumptionToDatabases(itemID);
 
     }
 
-    public Integer getItemConsumptionRate(Context context, Integer itemID){
+    public Integer getItemConsumptionRate(Context context, Long itemID){
 
         ConsumptionsDAO consumptionsDAO = new ConsumptionsDAO(context);
         return consumptionsDAO.getItemConsumptionRate(itemID);
